More of the same...

+++ GlkClient/GlkBufferedStream.m
@@ -63,8 +63,8 @@
 	}
 	
 	// Fill from high tide to the end of the buffer
-	int remaining = [fillData length];
-	int toCopy = remaining;
+	NSInteger remaining = [fillData length];
+	NSInteger toCopy = remaining;
 	
 	if (highTide + toCopy >= readAhead) toCopy = readAhead-highTide;
 	
@@ -151,11 +151,11 @@
 	return bytes[0];
 }
 
-- (bycopy NSString*) getLineWithLength: (int) maxLen {
+- (bycopy NSString*) getLineWithLength: (NSInteger) maxLen {
 	NSMutableString* res = [NSMutableString string];
 	
 	unichar ch;
-	int len = 0;
+	NSInteger len = 0;
 	do {
 		ch = [self getChar];
 		
@@ -173,7 +173,7 @@
 	return res;
  }
 
-- (bycopy NSData*) getBufferWithLength: (unsigned) length {
+- (bycopy NSData*) getBufferWithLength: (NSUInteger) length {
 	// Return nothing if there's nothing in the buffer and we can't fill it up
 	if (bufferRemaining == readAhead && ![self fillBuffer]) {
 		return nil;
@@ -181,11 +181,11 @@
 	
 	// Keep reading bytes until we run out of buffer
 	NSMutableData* result = [[[NSMutableData alloc] init] autorelease];
-	int toRead = length;
+	NSInteger toRead = length;
 	
 	while (bufferRemaining != readAhead && toRead > 0) {
 		// Work out how much to read this pass through
-		int thisPass = toRead;
+		NSInteger thisPass = toRead;
 		
 		if (lowTide + thisPass > readAhead) {
 			thisPass = readAhead - lowTide;
